Add client side and todo tests

+++ server/server.js
@@ -1,18 +1,17 @@
 var express = require('express');
 var bodyParser = require('body-parser');
 var { ObjectID } = require('mongodb');
-var { validId } = require('./utils/customFunctions');
-console.log(validId);
 var { mongoose } = require('./db/mongoose');
 var { Todo } = require('./models/todo');
 var { User } = require('./models/user');
 
 var app = express();
-const port = process.env.PORT || 3000;
+const port = process.env.PORT || 5000;
 
 app.use(bodyParser.json());
 
 app.post('/todos', (req, res) => {
+	console.log('at server', res);
 	var todo = new Todo({
 		text: req.body.text
 	});
@@ -41,10 +40,9 @@ app.get('/todos', (req, res) => {
 app.get('/todos/:id', (req, res) => {
 	var id = req.params.id;
 
-	validId(id);
-	// if (!ObjectID.isValid(id)) {
-	// 	return res.status(404).send();
-	// }
+	if (!ObjectID.isValid(id)) {
+		return res.status(404).send();
+	}
 
 	Todo.findById(id)
 		.then(todo => {
@@ -61,19 +59,18 @@ app.get('/todos/:id', (req, res) => {
 
 app.delete('/todos/:id', (req, res) => {
 	const { id } = req.params;
-	validId(id);
-	// if (!ObjectID.isValid(id)) {
-	// 	return res.status(404).send();
-	// }
+	if (!ObjectID.isValid(id)) {
+		return res.status(404).send();
+	}
 	Todo.findByIdAndDelete(id)
 		.then(todo => {
 			if (!todo) {
-				return res.status(400).send();
+				return res.status(404).send();
 			}
 			res.send(todo);
 		})
 		.catch(e => {
-			res.status(404).send();
+			res.status(400).send();
 		});
 });
